Refactor: move iced code out of game logic
Move code for loading card assets into its own file, under the
gui module. Important to "separate concerns". (Game logic belongs
in "game" code, gui code belongs in "gui" module.)
